new way to get info about client (JWT or password)

+++ src/services/client.service.js
@@ -1,15 +1,34 @@
-const { _router } = require('../app')
 const {Clients} = require('../models')
 const utils = require('./utils')
 const fs = require('fs')
 const path = require('path')
+const createError = require('http-errors')
 
 async function createClient(value) {
+    const chars = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+    let password = ""
+    for (i=0;i<8;i++) {
+        password += chars[utils.getRandomInt(0, chars.length)]
+    }
+    value.password = password
     const client = await Clients.create(value)
     return client
 } 
 
-async function getClientInfo(clientId) {
+async function getClientInfo(adminId, password, clientId) {
+    if (adminId) {
+        const admin = await utils.findAdmin(adminId)
+    } else {
+        const client = await Clients.findOne({
+            where: {
+                id: clientId,
+                password: password
+            }
+        })
+        if (!client) {
+            throw createError(404, "Username of password are false")
+        }
+    }
     const client = await utils.findClient(clientId)
     return client
 }
@@ -20,16 +39,26 @@ async function updateClient(adminid, clientId, value) {
 
     if (value.photo) {
         const photoPath = path.join(__dirname, "../public/", client.photo)
-        if (fs.existsSync(photoPath))
-        fs.unlinkSync(photoPath)
+        if (fs.existsSync(photoPath)) fs.unlinkSync(photoPath)
     }
     client.set(value)
     await client.save()
     return client
 }
 
+async function deleteClient(adminid, clientId) {
+    const admin = await utils.findAdmin(adminid)
+    const client = await utils.findClient(clientId)
+
+    const photoPath = path.join(__dirname, "../public/", client.photo)
+    if (fs.existsSync(photoPath)) fs.unlinkSync(photoPath)
+
+    await client.destroy()
+}
+
 module.exports = {
     createClient,
     getClientInfo,
-    updateClient
+    updateClient,
+    deleteClient
 }
